# server/docs/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: Feature flag system with rule-based evaluation, segment targeting,
    pack-based feature bundling, and A/B experiments.
  title: Feature Evaluator API
  version: "1.0"
paths:
...
  /admin/segments:
    get:
      description: Returns a paginated list of segments
      parameters:
        - description: Search filter
          in: query
          name: search
          schema:
            type: string
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
            default: 1
        - description: Page size
          in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Li\
                  stResponse-github_com_rendis_feature-evaluator_internal_dto_S\
                  egmentResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: List segments
      tags:
        - segments
    post:
      description: Creates a new segment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Cr\
                eateSegmentRequest"
        description: Segment creation payload
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Se\
                  gmentResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Create segment
      tags:
        - segments
  "/admin/segments/{key}":
    delete:
      description: Deletes a segment and all its records
      parameters:
        - description: Segment key
          in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Me\
                  ssageResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Delete segment
      tags:
        - segments
    get:
      description: Returns a single segment by key
      parameters:
        - description: Segment key
          in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Se\
                  gmentResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Get segment
      tags:
        - segments
    put:
      description: Updates an existing segment by key
      parameters:
        - description: Segment key
          in: path
          name: key
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Up\
                dateSegmentRequest"
        description: Segment update payload
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Se\
                  gmentResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Update segment
      tags:
        - segments
  "/admin/segments/{key}/data/import":
    post:
      description: Imports a full dataset version into a segment (replace mode)
      parameters:
        - description: Segment key
          in: path
          name: key
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Im\
                portSegmentDataRequest"
        description: Import payload
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Im\
                  portResultResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Import segment data
      tags:
        - segments
  "/admin/segments/{key}/records":
    get:
      description: Returns a paginated list of records for a segment
      parameters:
        - description: Segment key
          in: path
          name: key
          required: true
          schema:
            type: string
        - description: Search query
          in: query
          name: q
          schema:
            type: string
        - description: Page number
          in: query
          name: page
          schema:
            type: integer
            default: 1
        - description: Page size
          in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Li\
                  stResponse-github_com_rendis_feature-evaluator_internal_dto_S\
                  egmentRecordResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: List segment records
      tags:
        - segments
  "/admin/segments/{key}/schema":
    get:
      description: Returns the stored schema metadata for a segment
      parameters:
        - description: Segment key
          in: path
          name: key
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Se\
                  gmentSchemaResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_rendis_feature-evaluator_internal_dto.Er\
                  rorResponse"
      security:
        - BearerAuth: []
      summary: Get segment schema
      tags:
        - segments
...
servers:
  - url: /features
components:
  securitySchemes:
    ApiKeyAuth:
      description: API key for evaluation or admin access
      in: header
      name: X-Api-Key
      type: apiKey
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token
      in: header
      name: Authorization
      type: apiKey
    WorkspaceHeader:
      description: Workspace key for multi-tenant isolation
      in: header
      name: X-Workspace
      type: apiKey
...